requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-migrate>=4.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
gunicorn>=21.0.0
gevent>=23.0.0
msal>=1.24.0
python-dotenv>=1.0.0
twilio>=8.10.0
redis>=5.0.0
Werkzeug>=3.0.0
botbuilder-core>=4.17.1
